Add Telomare.Resolver module
This includes creating a telomare-resolver-tests suite and cleaning telomare-parser-tests
