+++ telomare.cabal
@@ -37,6 +37,7 @@ library
                      , Telomare.Optimizer
                      , Telomare.Parser
                      , Telomare.Possible
+                     , Telomare.Resolver
                      , Telomare.RunTime
                      , Telomare.TypeChecker
   build-depends:       base
@@ -108,7 +109,7 @@ test-suite telomare-test
                      , dlist
                      , hspec
                      , hspec-core
-		     , mtl
+                     , mtl
                      , telomare
                      , strict
                      , QuickCheck
@@ -131,11 +132,30 @@ test-suite telomare-parser-test
                      , megaparsec
                      , mtl
                      , recursion-schemes
-                     , Diff
-                     , pretty-show
                      , process
                      , lens
                      , unix
+  ghc-options:         -threaded -rtsopts -with-rtsopts=-N
+  default-language:    Haskell2010
+
+test-suite telomare-resolver-test
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      test
+  main-is:             ResolverTests.hs
+  other-modules:       Common
+  build-depends:       base
+                     , telomare
+                     , tasty
+                     , tasty-hunit
+                     , tasty-quickcheck
+                     , containers
+                     , QuickCheck
+                     , strict
+                     , megaparsec
+                     , unix
+                     , mtl
+                     , Diff
+                     , pretty-show
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N
   default-language:    Haskell2010
